Pulled out EvvoIsland component
* Created LocalEvvoIsland class for running locally, with no network or
  akka dependencies
* EvvoIsland renamed to EvvoIslandActor for clarity
* Integration tests changed to use LocalEvvoIsland
