+++ src/main/scala/com/evvo/island/IslandManager.scala
@@ -5,7 +5,7 @@ import java.io.File
 import akka.actor.{Actor, ActorSystem, Address, AddressFromURIString, Deploy, PoisonPill, Props}
 import akka.event.LoggingReceive
 import akka.remote.RemoteScope
-import com.evvo.island.EvvoIsland.{Emigrate, GetParetoFrontier, Run}
+import com.evvo.island.EvvoIslandActor.{Emigrate, GetParetoFrontier, Run}
 import com.evvo.island.population.{ParetoFrontier, TParetoFrontier, TScored}
 import com.typesafe.config.ConfigFactory
 
@@ -33,8 +33,6 @@ class IslandManager[Sol](val numIslands: Int,
     .withFallback(ConfigFactory.parseFile(new File("src/main/resources/application.conf")))
     .resolve()
 
-  println("making island manager")
-
   private val addresses: Vector[Address] = config.getList("nodes.locations")
     .unwrapped()
     .asScala.toVector
@@ -55,11 +53,9 @@ class IslandManager[Sol](val numIslands: Int,
   private val islands = (0 until numIslands).map(i => {
     val remoteChoice = i % addresses.length // pick round robin
     val address = addresses(remoteChoice)
-    system.actorOf(islandBuilder.props().withDeploy(Deploy(scope = RemoteScope(address))),
+    system.actorOf(islandBuilder.toProps.withDeploy(Deploy(scope = RemoteScope(address))),
       s"EvvoIsland${i}")
-  }).map(EvvoIsland.Wrapper[Sol])
-
-  println("islands created")
+  }).map(EvvoIslandActor.Wrapper[Sol])
 
   context.actorSelection("akka.tcp://RemoteEvvoNode@127.0.0.1:3451/user/Island0").resolveOne
 
@@ -84,45 +80,32 @@ class IslandManager[Sol](val numIslands: Int,
   : Future[Unit] = {
     Future {
       val runIslands = this.islands.map(_.runAsync(terminationCriteria))
-      println("RunIslands declared")
       runIslands.foreach(Await.result(_, Duration.Inf))
-      println("RunIslands foreach")
       this.finalParetoFrontier = Some(this.currentParetoFrontier())
-      println("got pareto frontier")
       this.islands.foreach(_.poisonPill())
-      println(s"this.finalParetoFrontier = ${this.finalParetoFrontier}")
       this.system.terminate()
-      println("finishing runBlocking in IslandManager")
     }
   }
 
   def runBlocking(terminationCriteria: TTerminationCriteria): Unit = {
     // TODO replace Duration.Inf
     Await.result(this.runAsync(terminationCriteria), Duration.Inf)
-    println("finishing runBlocking in IslandManager")
   }
 
   override def currentParetoFrontier(): TParetoFrontier[Sol] = {
-    println("IM CPF call")
     finalParetoFrontier match {
       case Some(paretoFrontier) =>
-        println("Some(paretoFrontier) ")
         paretoFrontier
       case None =>
-        println("None")
-        println(s"islands(0).currentParetoFrontier() = ${islands(0).currentParetoFrontier()}")
-
         val islandFrontiers = islands
           .map(_.currentParetoFrontier().solutions)
           .foldLeft(Set[TScored[Sol]]())(_ | _)
 
-        println("islandFrontiers")
         ParetoFrontier(islandFrontiers)
     }
   }
 
   override def poisonPill(): Unit = {
-    println("PoisonPill received")
     this.islands.foreach(_.poisonPill())
     self ! PoisonPill
   }
@@ -134,6 +117,6 @@ object IslandManager {
                 actorSystemName: String = "EvvoNode",
                 userConfig: String = "src/main/resources/application.conf")
                (implicit system: ActorSystem): TEvolutionaryProcess[Sol] = {
-    EvvoIsland.Wrapper(system.actorOf(Props(new IslandManager[Sol](numIslands, islandBuilder))))
+    EvvoIslandActor.Wrapper(system.actorOf(Props(new IslandManager[Sol](numIslands, islandBuilder))))
   }
 }
